refactor: Move the local clock check to a BlockHeader method
